+++ packages/unstyled/checkbox/src/Checkbox.tsx
@@ -12,6 +12,7 @@ import {
   composeEventHandlers,
 } from '@gluestack-ui/utils';
 import { useFormControlContext } from '@gluestack-ui/form-control';
+
 export const Checkbox = (StyledCheckbox: any) =>
   forwardRef(
     (
@@ -53,13 +54,15 @@ export const Checkbox = (StyledCheckbox: any) =>
 
       const _ref = React.useRef(null);
       const mergedRef = mergeRefs([ref, _ref]);
+      const ariaLabel =
+        combinedProps['aria-label'] || combinedProps.value || 'Checkbox';
 
       const { inputProps: groupItemInputProps } = checkboxGroupContext
         ? // eslint-disable-next-line react-hooks/rules-of-hooks
           useCheckboxGroupItem(
             {
               ...combinedProps,
-              'aria-label': combinedProps['aria-label'],
+              'aria-label': ariaLabel,
               'value': combinedProps.value,
             },
             checkboxGroupContext.state,
@@ -70,7 +73,7 @@ export const Checkbox = (StyledCheckbox: any) =>
           useCheckbox(
             {
               ...combinedProps,
-              'aria-label': combinedProps['aria-label'],
+              'aria-label': ariaLabel,
             },
             state,
             //@ts-ignore