+++ packages/unstyled/accordion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gluestack-ui/accordion
 
+## 1.0.6
+
+### Patch Changes
+
+- onpress issue on accordion trigger
+
 ## 1.0.5
 
 ### Patch Changes